$PSCmdlet.ShouldProcess() implemented for all system changes, medium ConfirmImpact to prevent undesirable interruptions

+++ PSSpeedTest/private/Remove-iPerf3Port.ps1
@@ -11,14 +11,16 @@
 #>
 
 function Remove-iPerf3Port {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess=$True,ConfirmImpact="Medium")]
     Param()
 
     Write-Verbose -Message "Removing inbound and outbound iperf3 firewall rules."
 
     try {
-        Get-NetFirewallRule -DisplayName "iPerf3 Server Inbound TCP Rule" | Remove-NetFirewallRule
-        Get-NetFirewallRule -DisplayName "iPerf3 Server Outbound TCP Rule" | Remove-NetFirewallRule
+        if ($PSCmdlet.ShouldProcess("iperf3 Inbound/Outbound Firewall Rules", "Remove-NetFirewallRule")) {
+            Get-NetFirewallRule -DisplayName "iPerf3 Server Inbound TCP Rule" | Remove-NetFirewallRule
+            Get-NetFirewallRule -DisplayName "iPerf3 Server Outbound TCP Rule" | Remove-NetFirewallRule
+        }
     }
     catch {
         Write-Verbose -Message "Firewall rules not found - no action taken."
